% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haploMatch.R
\name{haploMatch}
\alias{haploMatch}
\title{Match haplotype}
\usage{
haploMatch(haploStrA, haploStrB)
}
\arguments{
\item{haploStrA}{Matrix with given haplotype structure.
\code{nrow(haploStrA)} is number of SNP locations and \code{ncol(haploStrA)}
is number of estimated haploypes.}

\item{haploStrB}{Matrix of a candidate list of haplotypes.
\code{nrow(haploStrB)} is number of SNP locations and \code{ncol(haploStrB)}
is number of candidate haploypes.}
}
\value{
Integer vector with indexes of haploStrB which fit haploStrA best.
}
\description{
For a set of given haplotypes finds best matching haplotypes from a
canditate list with loss function being the number of non equal elements.
}
\examples{
 Compare the reconstruction with 3 dominant haplotypes to the one with 7 dominant haplotypes
 data(ExampleDataset)
 reconstruction1 <- haploSep(data = Y, nHaplo = 3, stabEval = FALSE, bias = TRUE)
 reconstruction2 <- haploSep(data = Y, nHaplo = 7, stabEval = FALSE, bias = TRUE)
 haploMatch(reconstruction1$haploStr, reconstruction2$haploStr)
 
 
}
