% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haploSelect.R
\name{haploSelect}
\alias{haploSelect}
\title{Model selector for haplotype separation}
\usage{
haploSelect(data)
}
\arguments{
\item{data}{Numeric observation matrix with nrow(data) being the number of
SNP locations and ncol(data) the number of samples (e.g. time points).}
}
\value{
integer. The estimated number of essential haplotypes.
}
\description{
For an observed data set of allele frequencies finds the 
estimated number of dominant haplotypes given the 
observed allele frequency changes.
}
\examples{
data(ExampleDataset)

haploSelect(data = Y)

}
