% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haploPlot.R
\name{plot.haplo}
\alias{plot.haplo}
\title{Plot the reconstruction of haplotypes}
\usage{
\method{plot}{haplo}(x, type = c("both", "str", "frq"), compare = NULL, ...)
}
\arguments{
\item{x}{an object from class haplo}

\item{type}{character string giving the type of plot desired
\itemize{
\item
"\code{str}" plot the structure of haplotypes
\item
"\code{frq}" plot the frequency of haplotypes
\item
"\code{both}" plot both the structure and the frequency of haplotypes
}}

\item{compare}{an object from class haplo to be compared with x}

\item{...}{Arguments to be passed to methods, such as \link{graphical parameters} (see \code{\link[graphics]{par}}).}
}
\description{
Function for plotting objects from class haplo
}
\examples{
 data(ExampleDataset)
 m <- haploSelect(data = Y)
 reconstruction <- haploSep(data = Y, nHaplo = m, stabEval = TRUE, bias = TRUE)
 
 plot.haplo(reconstruction, type = "both")


}
